api-common done + pagination without html & css
