+++ src/sass/index.scss
@@ -1,9 +1,9 @@
 @import './base/common';
 @import './base/global';
-@import './base/pagination';
 
 @import './utils/mixins';
 @import './utils/variables';
 @import './utils/placeholders';
+@import './utils/pagination';
 
 @import './layouts/header'
